Make sure post categories is loaded as Array
Categories can be set in various ways, but we will go for
the name 'categories' and a yaml list ( Array ) always.
We make sure if category is used, we convert it.

This commit also fixes a bug that would discard other settings
than those we specified per default. Utterson should ofcourse
always keep wierd/custom settings, regardsless if it uses them.