Rewrite Post.all to load array of posts
Instead of some arbitrary array of contents.
It introduces another problem though. We load the entire posts
content into an array, we should create some sort of lazy loading
of the blog post itself ( which is by far the biggest and not at all
relevant for 99% of all use cases of Post.all )